Divide Animator and movement
