+++ 2D Homework/Assets/Tasks/Platformer/Scripts/PlayerAnimator.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9ece2d1a609344729d7cf797fe021c23
+timeCreated: 1702112334
